fix: move Support development back to Self-hosted tab
About info (version/source/license) stays on Advanced (universal).
Support development (Ko-fi) gates to Self-hosted tab so future
hosted-tier paying customers don't see a donation prompt.
