+++ main.js
@@ -2174,14 +2174,6 @@ secret.md`).setValue(plugin.settings.ignorePatterns).onChange(async (value) => {
     text: "github.com/Rasbandit/Engram-obsidian-sync",
     href: "https://github.com/Rasbandit/Engram-obsidian-sync"
   }), aboutList.createEl("li").createSpan({ text: "License: MIT" });
-  let kofiLink = new import_obsidian10.Setting(containerEl).setName("Support development").setDesc(
-    "If this plugin saves you time, consider supporting development. Optional and appreciated."
-  ).controlEl.createEl("a", {
-    cls: "engram-kofi-button",
-    href: "https://ko-fi.com/rasbandit",
-    attr: { target: "_blank", rel: "noopener" }
-  }), iconSpan = kofiLink.createSpan({ cls: "engram-kofi-icon" });
-  (0, import_obsidian10.setIcon)(iconSpan, "coffee"), kofiLink.createSpan({ text: "Support on Ko-fi" });
 }
 function renderIgnoreWarnings(containerEl, app, plugin, redisplay) {
   let currentIgnores = plugin.settings.ignorePatterns, detected = [];
@@ -2268,7 +2260,15 @@ function renderSelfHostedTab(ctx) {
         value && value !== plugin.settings.vaultId && (plugin.settings.vaultId = value, plugin.api.setVaultId(value), await plugin.saveSettings(), redisplay());
       });
     });
-  }));
+  })), new import_obsidian11.Setting(containerEl).setName("Support development").setHeading();
+  let kofiLink = new import_obsidian11.Setting(containerEl).setDesc(
+    "If this plugin saves you time, consider supporting development. Optional and appreciated."
+  ).controlEl.createEl("a", {
+    cls: "engram-kofi-button",
+    href: "https://ko-fi.com/rasbandit",
+    attr: { target: "_blank", rel: "noopener" }
+  }), iconSpan = kofiLink.createSpan({ cls: "engram-kofi-icon" });
+  (0, import_obsidian11.setIcon)(iconSpan, "coffee"), kofiLink.createSpan({ text: "Support on Ko-fi" });
 }
 
 // src/settings.ts
