+++ src/tabs/advanced-tab.ts
@@ -1,4 +1,4 @@
-import { Notice, Setting, TFolder, setIcon } from "obsidian";
+import { Notice, Setting, TFolder } from "obsidian";
 import { PreSyncModal, WipeConfirmModal } from "../pre-sync-modal";
 import type { TabContext } from "./types";
 
@@ -217,21 +217,6 @@ export function renderAdvancedTab(ctx: TabContext): void {
 
 	const licenseItem = aboutList.createEl("li");
 	licenseItem.createSpan({ text: "License: MIT" });
-
-	const supportSetting = new Setting(containerEl)
-		.setName("Support development")
-		.setDesc(
-			"If this plugin saves you time, consider supporting development. Optional and appreciated.",
-		);
-
-	const kofiLink = supportSetting.controlEl.createEl("a", {
-		cls: "engram-kofi-button",
-		href: "https://ko-fi.com/rasbandit",
-		attr: { target: "_blank", rel: "noopener" },
-	});
-	const iconSpan = kofiLink.createSpan({ cls: "engram-kofi-icon" });
-	setIcon(iconSpan, "coffee");
-	kofiLink.createSpan({ text: "Support on Ko-fi" });
 }
 
 /** Scan vault for problematic directories and render warnings with add-to-ignore buttons. */
